requirements.txt
pbr>4.0.3 # Apache-2.0
pyquil>=2.2.1 # Apache-2.0
qiskit-terra<0.8.0,>=0.7.0 # Apache-2.0
